added skeleton for edges hierarchy

+++ graph.h
@@ -10,18 +10,16 @@ enum class GraphRepresentation
 };
 
 
-
 class IGraph
 {
 public:
-	virtual int FindMinimumCut() const = 0;
+	virtual size_t FindMinimumCut() const = 0;
 
 };
 
-//TODO: vertex and edge should have each one classes in order to satisty whole class hierarchy polymorphic behaviour
-//edges shall be adopted according to graph
 
 
+//TODO: implement DFS
 //this will be the very basic grapsh representation
 class Graph: public IGraph
 {
@@ -29,7 +27,7 @@ public:
     Graph(const std::vector<std::vector<int>> &vAdjMatrix);
     void AddEdge(const std::pair<int, int> &edge);
     void AddVertex(const std::vector<int> &vVertex);
-    int FindMinimumCut() const;
+    size_t FindMinimumCut() const;
     int FindMinimumPath(int iNode1, int iNode2) const;
 private:
     void BuildFromMatrix(const std::vector<std::vector<int>> &vAdjMatrix);
